Time: 65 ms (5.02%) | Memory: 115.9 MB (38.35%) - LeetSync

+++ 435-non-overlapping-intervals/non-overlapping-intervals.java
@@ -1,6 +1,6 @@
 class Solution {
     public int eraseOverlapIntervals(int[][] intervals) {
-        Arrays.sort(intervals, Comparator.comparingInt(o -> o[1]));
+        Arrays.sort(intervals, Comparator.comparingDouble(o -> o[1]));
         int count = 0;
         int lastEnd = intervals[0][1];
         for (int i = 1; i < intervals.length; i++) {
